feat: esmfy
Move js_obj_to_alist to pair.js
